box: move space_cache to a separate file
I'm going to extend space cache API so it should be separated.
One function went to space.h/c.
No logical changes.

NO_DOC=refactoring
NO_CHANGELOG=refactoring
NO_TEST=refactoring
