+++ src/box/CMakeLists.txt
@@ -167,6 +167,7 @@ set(box_sources
     vy_quota.c
     request.c
     space.c
+    space_cache.c
     space_def.c
     sequence.c
     ck_constraint.c
