+++ src/box/schema.cc
@@ -30,7 +30,6 @@
  */
 #include "schema.h"
 #include "sequence.h"
-#include "tuple.h"
 #include "assoc.h"
 #include "alter.h"
 #include "scoped_guard.h"
@@ -55,19 +54,10 @@
  * it.
  */
 
-/** All existing spaces. */
-static struct mh_i32ptr_t *spaces;
-static struct mh_strnptr_t *spaces_by_name;
 static struct mh_i32ptr_t *sequences;
 /** Public change counter. On its update clients need to fetch
  *  new space data from the instance. */
 uint32_t schema_version = 0;
-/**
- * Internal change counter. Grows faster, than the public one,
- * because we need to remember when to update pointers to already
- * non-existent space objects on space:truncate() operation.
- */
-uint32_t space_cache_version = 0;
 
 /** Persistent version of the schema, stored in _schema["version"]. */
 uint32_t dd_version_id = 0;
@@ -79,178 +69,11 @@ struct rlist on_alter_func = RLIST_HEAD_INITIALIZER(on_alter_func);
 
 struct entity_access entity_access;
 
-bool
-space_is_system(struct space *space)
-{
-	return space->def->id > BOX_SYSTEM_ID_MIN &&
-		space->def->id < BOX_SYSTEM_ID_MAX;
-}
-
-/** Return space by its number */
-struct space *
-space_by_id(uint32_t id)
-{
-	mh_int_t space = mh_i32ptr_find(spaces, id, NULL);
-	if (space == mh_end(spaces))
-		return NULL;
-	return (struct space *) mh_i32ptr_node(spaces, space)->val;
-}
-
-/** Return space by its name */
-struct space *
-space_by_name(const char *name)
-{
-	mh_int_t space = mh_strnptr_find_str(spaces_by_name, name,
-					     strlen(name));
-	if (space == mh_end(spaces_by_name))
-		return NULL;
-	return (struct space *)mh_strnptr_node(spaces_by_name, space)->val;
-}
-
 /** Return current schema version */
 uint32_t
 box_schema_version(void)
 {
 	return schema_version;
-}
-
-/**
- * Visit all spaces and apply 'func'.
- */
-int
-space_foreach(int (*func)(struct space *sp, void *udata), void *udata)
-{
-	mh_int_t i;
-	struct space *space;
-	char key[6];
-	assert(mp_sizeof_uint(BOX_SYSTEM_ID_MIN) <= sizeof(key));
-	mp_encode_uint(key, BOX_SYSTEM_ID_MIN);
-
-	/*
-	 * Make sure we always visit system spaces first,
-	 * in order from lowest space id to the highest..
-	 * This is essential for correctly recovery from the
-	 * snapshot, and harmless otherwise.
-	 */
-	space = space_by_id(BOX_SPACE_ID);
-	struct index *pk = space ? space_index(space, 0) : NULL;
-	if (pk) {
-		struct iterator *it = index_create_iterator(pk, ITER_GE,
-							    key, 1);
-		if (it == NULL)
-			return -1;
-		int rc;
-		struct tuple *tuple;
-		while ((rc = iterator_next(it, &tuple)) == 0 && tuple != NULL) {
-			uint32_t id;
-			if (tuple_field_u32(tuple, BOX_SPACE_FIELD_ID, &id) != 0)
-				continue;
-			space = space_cache_find(id);
-			if (space == NULL)
-				continue;
-			if (! space_is_system(space))
-				break;
-			rc = func(space, udata);
-			if (rc != 0)
-				break;
-		}
-		iterator_delete(it);
-		if (rc != 0)
-			return -1;
-	}
-
-	mh_foreach(spaces, i) {
-		space = (struct space *) mh_i32ptr_node(spaces, i)->val;
-		if (space_is_system(space))
-			continue;
-		if (func(space, udata) != 0)
-			return -1;
-	}
-
-	return 0;
-}
-
-void
-space_cache_replace(struct space *old_space, struct space *new_space)
-{
-	assert(new_space != NULL || old_space != NULL);
-	if (new_space != NULL) {
-		/*
-		 * If the replaced space has a different name, we
-		 * must explicitly delete it from @spaces_by_name
-		 * cache. Note, since a space id never changes, we
-		 * don't need to do so for @spaces cache.
-		 */
-		struct space *old_space_by_name = NULL;
-		if (old_space != NULL && strcmp(space_name(old_space),
-						space_name(new_space)) != 0) {
-			const char *name = space_name(old_space);
-			mh_int_t k = mh_strnptr_find_str(spaces_by_name, name,
-							 strlen(name));
-			assert(k != mh_end(spaces_by_name));
-			old_space_by_name = (struct space *)
-				mh_strnptr_node(spaces_by_name, k)->val;
-			mh_strnptr_del(spaces_by_name, k, NULL);
-		}
-		/*
-		 * Insert @new_space into @spaces cache, replacing
-		 * @old_space if it's not NULL.
-		 */
-		const struct mh_i32ptr_node_t node_p = { space_id(new_space),
-							 new_space };
-		struct mh_i32ptr_node_t old, *p_old = &old;
-		mh_i32ptr_put(spaces, &node_p, &p_old, NULL);
-		struct space *old_space_by_id = p_old != NULL ?
-				(struct space *)p_old->val : NULL;
-		assert(old_space_by_id == old_space);
-		(void)old_space_by_id;
-		/*
-		 * Insert @new_space into @spaces_by_name cache.
-		 */
-		const char *name = space_name(new_space);
-		uint32_t name_len = strlen(name);
-		uint32_t name_hash = mh_strn_hash(name, name_len);
-		const struct mh_strnptr_node_t node_s = { name, name_len,
-							  name_hash, new_space };
-		struct mh_strnptr_node_t old_s, *p_old_s = &old_s;
-		mh_strnptr_put(spaces_by_name, &node_s, &p_old_s, NULL);
-		if (old_space_by_name == NULL && p_old_s != NULL)
-			old_space_by_name = (struct space *)p_old_s->val;
-		assert(old_space_by_name == old_space);
-		(void)old_space_by_name;
-	} else {
-		/*
-		 * Delete @old_space from @spaces cache.
-		 */
-		mh_int_t k = mh_i32ptr_find(spaces, space_id(old_space), NULL);
-		assert(k != mh_end(spaces));
-		struct space *old_space_by_id =
-			(struct space *)mh_i32ptr_node(spaces, k)->val;
-		assert(old_space_by_id == old_space);
-		(void)old_space_by_id;
-		mh_i32ptr_del(spaces, k, NULL);
-		/*
-		 * Delete @old_space from @spaces_by_name cache.
-		 */
-		const char *name = space_name(old_space);
-		k = mh_strnptr_find_str(spaces_by_name, name, strlen(name));
-		assert(k != mh_end(spaces_by_name));
-		struct space *old_space_by_name =
-			(struct space *)mh_strnptr_node(spaces_by_name, k)->val;
-		assert(old_space_by_name == old_space);
-		(void)old_space_by_name;
-		mh_strnptr_del(spaces_by_name, k, NULL);
-	}
-	space_cache_version++;
-
-	if (trigger_run(&on_alter_space, new_space != NULL ?
-					 new_space : old_space) != 0) {
-		diag_log();
-		panic("Can't update space cache");
-	}
-
-	if (old_space != NULL)
-		space_invalidate(old_space);
 }
 
 static int
@@ -378,8 +201,7 @@ schema_init(void)
 		key_parts[i] = key_part_def_default;
 
 	/* Initialize the space cache. */
-	spaces = mh_i32ptr_new();
-	spaces_by_name = mh_strnptr_new();
+	space_cache_init();
 	func_cache_init();
 	sequences = mh_i32ptr_new();
 	/*
@@ -521,18 +343,7 @@ schema_init(void)
 void
 schema_free(void)
 {
-	if (spaces == NULL)
-		return;
-	while (mh_size(spaces) > 0) {
-		mh_int_t i = mh_first(spaces);
-
-		struct space *space = (struct space *)
-				mh_i32ptr_node(spaces, i)->val;
-		space_cache_replace(space, NULL);
-		space_delete(space);
-	}
-	mh_i32ptr_delete(spaces);
-	mh_strnptr_delete(spaces_by_name);
+	space_cache_destroy();
 	func_cache_destroy();
 
 	while (mh_size(sequences) > 0) {
