+++ src/box/space.c
@@ -117,6 +117,13 @@ space_fill_index_map(struct space *space)
 	}
 }
 
+bool
+space_is_system(struct space *space)
+{
+	return space->def->id > BOX_SYSTEM_ID_MIN &&
+	       space->def->id < BOX_SYSTEM_ID_MAX;
+}
+
 /**
  * Initialize constraints that are defined in @a space format.
  * Can return nonzero in case of error (diag is set).
